rril: fix bad data size in QUERY_AVAILABLE_NETWORKS reply
BZ: 209565

The data size of the QUERY_AVAILABLE_NETWORKS reply was set by
counting the number of "(" in the COPS? reply. If an operator
had a "(" in its name, this size did not match the number of
filled entries in the array leading to an exception in RILJ (or
a potential segmentation fault in the serialization code in
rild).

Fix the issue by aligning the data size with the actual number
of OPINFO structure filled after having parsed the modem reply.

Change-Id: Ic3ce9a508e85f84580833bfdc4acb09443ba829f
Signed-off-by: Lionel Ulmer <lionel.ulmer@intel.com>