rril: Update MAX_OP_NAME_NUM to 6
BZ: 210763

The MNC can have two or three digits. This is currently
not take into account on the RRIL where we assume that
the MNC is coded with only two digits.

This patch updates the MAX_OP_NAME_NUM constant to take
into account the network where the MNC is coded with
three digits.

Change-Id: Ied55dd06b98d41940f97a632497567319cdb93ce
Signed-off-by: Guillaume Lucas <guillaume.lucas@intel.com>