rril: add IM_CN_Signaling flag to CGDCONT
BZ: 211387

When the data profile is IMS, set the IM CN Signalling flag along
with the request pcscf flag.
At RIL level, we don't know if a setup data call request is done
for an APN that supports multiple types.
In order to allow RIL to know if APN supports multiple types, the
dataProfile parameter is changed to a bitmask. If the bit at index
RIL_DATA_PROFILE_XXX is set, it means that APN supports this type.

Change-Id: I1f0c58e4ef940dd908ec48ca2509475d9120681a
Signed-off-by: Alexandre d'Alton <alexandre.d'alton@intel.com>