rril: suppress XMCI sending on serving cell change
BZ: 212719

As XCELLINFO took up to 5 seconds to be answered by the modem,
values were pre-cached each time the AP detected that the UE
changed serving cell so as to be able to send values back to the
framework as fast as possible (i.e. cached values were sent
instead of sending XCELLINFO to the modem and waiting for an
answer).

With XMCI being immediate, this sending of the cached values has
been suppressed meaning that there is no need anymore to fill
the cache at cell change.

This change configures this behavior in the modem classes
keeping cache fill on 6260 / 6360 modems and removing it on
7160 / 7260.

Change-Id: Ibaee8474a6bba159a1ddaa0ff4cf58354001cc02
Signed-off-by: Lionel Ulmer <lionel.ulmer@intel.com>