rril: fix the handling of the subscription id
With the Multi SIM introduction in the L dessert, we need to fix
the way we are handling the subscription id.

The rril will now rely on the CONF files to identify if we are
in a multi SIM (DSDS) or multi modem (DSDA) configuration.

Several assumptions need to be taken into account:
- L dessert allows us to create different ril-daemon services and
run them in //. One ril-daemon service handles one SIM.

- rild handles a new option passed as an argument: "-c" option
for identifying the client id

- The first and default service is called "ril-daemon" and it is
started in the init.rc file. No changes are allowed in this file.
There is no "-c" option passed in parameter so in the rril, the
value of the client id will be 0

- the second ril-daemon service is called "ril-daemon1" and it is
created in the init.modem.rc file. This service will have the value
"2" for the client id because it needs to be connected to the
socket "rild2" and the rild implementation takes the -c arg to
find the socket id.

Change-Id: I44f50454a0ea4fa10c97cfecf53a7bbe1e686854
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-833
Signed-off-by: Frederic Predon <frederic.predon@intel.com>
Reviewed-on: https://android.intel.com:443/236094