rapid_ril: delete data channel port, when modem doesn't support data
The DLC from 9 to 13 are used by data port. They depend on "DataCapable"
parameter which is not set on 2230 modem. Therefore the result of init
can be ignored. As gsmtty81 is used for modem tracing only, the count
of DLC will not reach the correct value, the ril init will fail. Even if
it fails, ril shall still set up

Change-Id: I4880bdc9f71b0a812a6a0045cc4d5ffba1b4d813
Signed-off-by: zhou.yu <zhou.yu@intel.com>
Reviewed-on: https://android.intel.com:443/247519
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-5106
Reviewed-on: https://android.intel.com:443/287413