Fix data call setup failure seen with CMW500
If the APN is stored with APN type set to *, then
framework triggers SETUP_DATA_CALL with dataProfile
set to all profiles. This results in AT+CGDCONT issued
with emergency_indication set to 1. As a result,
pdp context activation fails with CME ERROR: 148.

Fix the issue by setting the emergency_indication to
true if the SETUP_DATA_CALL is meant only for emergency
call.

Change-Id: Iafc3828f1e487b7ef39bc6b94ca5fa1ad9f3d7a4
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-3696
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Reviewed-on: https://android.intel.com:443/281299