Fix issue in accepting call when there is an active and held call
When there is an an active and held call, accepting an incoming
call results in framework sending the ril request
RIL_REQUEST_SWITCH_WAITING_OR_HOLDING_AND_ACTIVE. As rapid ril
sends wrong command, active call is released but incoming call is
not answered.

When there is an active and held call, send the command to
accept the MT call based on active, held and waiting call status.

Change-Id: I5cf92b2559820c3ccac4f76756430f9b8d6aef18
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-2325
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Reviewed-on: https://android.intel.com:443/281256