Fix for not able to disconnect the call in U8 state
Call disconnection on U8 state is not supported. Due to this,
3GPP conformance test case 26.8.1.3.5.3 fails.

When an incoming call is answered by user but not acknowledged
by network, ending of a call results in wrong ril request
(RIL_REQUEST_HANGUP_WAITING_OR_BACKGROUND) sent. Upon receiving
this ril request, rapid ril sends AT+CHLD=0 which fails with
CME ERROR: 4. Due to this call is never ended.

Fix the issue by sending AT+CHLD=1<cid> to release the incoming
call which is answered but not yet acknowledged by network.

Change-Id: I8c911aa23e609675e539e269caa53cb67b4f9b6e
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-4316
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Reviewed-on: https://android.intel.com:443/285225