rril: Remove use of RIL_E_INVALID_PARAMETER
As RIL_E_INVALID_PARAMETER is not provided by Google RIL,
usage of this error code needs to be removed from vendor ril.
Also, as initial analysis shows that XEER doesn't need to be
sent to get extended error report, sending of XEER command
and handling of XEER response is removed.

Change-Id: If416d6a70a61943ba9783b90472a6055844d34b7
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-5649
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Reviewed-on: https://android.intel.com:443/289624