Fix for "No service" shown instead of "Emergency calls only"
When the registration state is denied and emergency calls are
allowed, device shows emergeny calls only in the UI. But, this
text changes from "Emergency calls only" to "No service" when
the user toggles the screen state.

When the screen state changes from off to on, rapid ril queries
the registration state from modem. As modem is searching for
network and since it is a registration state query, modem returns
not registered but searching which is an intermediate state.
As rapid ril reports this state to framework, framework will
display "No service".

Fix the issue by not storing the intermediate state in rril.
So, when the modem returns intermediate state on registration
state query, previous cached registration state will be completed
to framework. Also, report registration state as emergency calls
possible if the device supports voice and emergency calls possible
case is satisfied.

Change-Id: If342a63eaa5a2c71132db587de780b789d76ae0a
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-5808
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Reviewed-on: https://android.intel.com:443/290109