Fix device LTE deregistering from network on data reactivation
When user deactivates and reactivates the data in LTE,
device deregisters from the network.

Upon user deactivating the data in LTE, interface is removed
and also address information is reset. So, when the user
reactivates data, rapid ril is not able to bring the interface
up due to all address information cleared out. This results in
AT+CGACT=0,1 sent to modem which results in device deregistering
from network and reregistering again.

Delete all address information only when data call is
reset due to force deactivation(power off) or deactivation
which results in sending of deactivation command to modem.

Change-Id: I6581653fe3adeda8dc4e5caced130b83db58dd15
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-5834
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Reviewed-on: https://android.intel.com:443/290239