rril: send AT+CEMODE as part of init string
CEMODE was sent at the same time as the XACT command when
setting the preferred network type (this was done to be able
to handle the - now unsupported - special case of LTE-only).

This caused issue in the AT&T variant where the
REQUEST_SET_PREFERRED_NETWORK_TYPE API is filtered at framework
side leading to CEMODE never being sent and so the mode of
operation remained 1 (the default value) even on a data only
configuration.

Fix the issue by setting the mode of operation based upon the
data only / data+voice configuration as part of the ATCMD
channel init string for all modems inheriting from 7160 modem.

Change-Id: I2966168fd7ad229cff256e26ad9dbba49da68825
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-6635
Signed-off-by: Lionel Ulmer <lionel.ulmer@intel.com>
Reviewed-on: https://android.intel.com:443/294597