rapid_ril: Remove encryption condition
With new L-dessert requirement, the data partition
is encrypted by default wihtout password. Currently,
the rril is checking is the DUT is decrypted or not
and send the "XSIMSTATE=1" AT command only when the
accurate property is set, which is not the case when
no encryption password is set.

Checking the decrypt state should not be used during
the RIL initialization as the value depends on the
partition state. It could be default encrypted or
starting the minimum framework or restarting main
components, etc..
We should just check this state before acquire the
modem resource or not to avoid registering on the
network during the 1st phase of the telephony stack
boot when the device is encrypted with a specific
password.

Change-Id: I553a7051c039ca6785437cec0906966457af5dd5
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-7125
Signed-off-by: Nicolas LOUIS <nicolasx.louis@intel.com>
Reviewed-on: https://android.intel.com:443/295063