Fix device losing CS when disabling data
when device is booted with preferred network type set to 3G,
changing of preferred network mode to 4G and then disabling
data results in loss of CS network.

When a data connection is established in 3G and handed over
to 4G, disabling of data results in deactivation of active
data connection. As deactivation of last data connection will
result in device deregistering from network and reregistering
again deactivation of last data connection should not be done
in 4G.

Following changes are done to fix the issue:
- First requested connection cid as default pdn cid if there is
no previous default PDN cid set.
- If the deactivation request is for default PDN and if its
interface can be stopped(device is not EPS registered or can be
force cleaned up) then the next active connection's cid is set as
default PDN cid or the default PDN context id is set to 0.
- Return true if one of the Apn's provided to IsApnEqual is empty
or NULL.

Change-Id: I271f690e9e640a63501832dd2e40b4999a85b8d0
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-13131
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Reviewed-on: https://android.intel.com:443/310236