Fix device not registering on flight mode exit
With FMMO disabled, device is not registering to network
on flight mode deactivation.

When there is no change in sim state, framework does't restore
the network selection mode. So, network selection command
AT+COPS=<> is not sent to modem if rapid ril doesn't restore
the network selection mode. Rapid ril tries to restore the
network selection mode but it doesn't restore due to empty
initial attach apn. As initial attach apn is not stored for
non-LTE modems, it will be empty. So, validity of initial attach
apn for non-LTE modems should not be done.

Following changes done to fix the issue:
- Move initial attach apn parameters validity to modem class
which supports LTE.
- Pass the next state to SetInitialAttachApn function to make
sure next command is sent based on next state.

Change-Id: I02fc94c1ce320d3fac3b2f6428d35bd92f25e271
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-18246
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Reviewed-on: https://android.intel.com:443/322272