Fix device not registering on flight mode exit on certain cases
If modem is powered on during flight mode(by Wifi or any modem
clients), device is not registering on flight mode as the
network selection is not restored on flight mode exit.

If modem is powered on during flight modem, framework sends
the network selection request but rapid ril has rejected it
as radio is not powered on. On flight mode exit, framework
won't send the network selection request as there is no
change in sim state.

Fix the issue by caching the network selection mode request
in rapid ril if radio is not yet powered on and restore it
once radio is powered on and sim is ready.

Change-Id: I81a49fc773965a65ce6686376625077374a16703
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-29346
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Reviewed-on: https://android.intel.com:443/363465